fix(workspace): resolve remaining test failures in focus-ir + focus-rules + focus-demo-seed

+++ crates/focus-storage/src/sqlite/event_dedup.rs
@@ -75,7 +75,7 @@ impl EventDeduplicator for SqliteAdapter {
                 )
                 .map_err(|e| DedupeError::DatabaseError(format!("purge: {e}")))?;
 
-            Ok(count as usize)
+            Ok(count)
         })
         .await
         .map_err(|e| DedupeError::DatabaseError(format!("purge_older_than join: {e}")))?
